Use parent image from fedora registry
Ideally, we could just use the image from registry.redhat.io/rhel8/go-toolset:latest
But, the registry requires authentication and there doesn't seem to be
away to provide credentials for the builds done in quay.io.

The next best thing is to use the golang image from docker hub, docker.io/golang:1.16
You guessed it, that also requires authentication to avoid rate
limiting.

The third option is to use a Fedora base image and install golang on it.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>